+++ src/Entity.h
@@ -16,10 +16,10 @@ class LivingEntity : public Entity {
 	public:
 	Level* level;
 	uint64_t posX, posY;
-	float health, maxHealth;
+	int health, maxHealth;
 	bool friendly;
 	std::string name;
 	LivingEntity(Level* level, uint64_t posX, uint64_t posY);
-	bool hurt(float damage);
-	float heal(float value);
+	bool hurt(int damage);
+	int heal(int value);
 };
